+++ app/code/Magento/Catalog/Controller/Adminhtml/Product/Initialization/StockDataFilter.php
@@ -56,7 +56,7 @@ class StockDataFilter
         if ($stockData['use_config_manage_stock'] == 1 && !isset($stockData['manage_stock'])) {
             $stockData['manage_stock'] = $this->stockConfiguration->getManageStock();
         }
-        if (isset($stockData['qty']) && (double)$stockData['qty'] > self::MAX_QTY_VALUE) {
+        if (isset($stockData['qty']) && (float)$stockData['qty'] > self::MAX_QTY_VALUE) {
             $stockData['qty'] = self::MAX_QTY_VALUE;
         }
