+++ app/code/Magento/Catalog/Model/Layer/Filter/DataProvider/Price.php
@@ -215,7 +215,7 @@ class Price
                         $index++;
                     } while ($range > self::MIN_RANGE_POWER && count($items) < 2);
                 } else {
-                    $range = (double)$this->getRangeStepValue();
+                    $range = (float)$this->getRangeStepValue();
                 }
             }
 
@@ -310,7 +310,7 @@ class Price
             return false;
         }
         foreach ($filter as $v) {
-            if ($v !== '' && $v !== '0' && (double)$v <= 0 || is_infinite((double)$v)) {
+            if ($v !== '' && $v !== '0' && (float)$v <= 0 || is_infinite((float)$v)) {
                 return false;
             }
         }
