+++ app/code/Magento/Checkout/Model/Cart.php
@@ -528,7 +528,7 @@ class Cart extends DataObject implements CartInterface
                 continue;
             }
 
-            $qty = isset($itemInfo['qty']) ? (double)$itemInfo['qty'] : false;
+            $qty = isset($itemInfo['qty']) ? (float)$itemInfo['qty'] : false;
             if ($qty > 0) {
                 $item->setQty($qty);
