+++ app/code/Magento/Paypal/Model/ResourceModel/Report/Settlement.php
@@ -77,7 +77,7 @@ class Settlement extends \Magento\Framework\Model\ResourceModel\Db\AbstractDb
                     /*
                      * Converting numeric
                      */
-                    $rows[$key]['fee_amount'] = (double)$rows[$key]['fee_amount'];
+                    $rows[$key]['fee_amount'] = (float)$rows[$key]['fee_amount'];
                     /*
                      * Setting reportId
                      */
