+++ app/code/Magento/Quote/Model/Quote/Item/AbstractItem.php
@@ -467,7 +467,7 @@ abstract class AbstractItem extends \Magento\Framework\Model\AbstractExtensibleM
     {
         if (!$this->hasBaseCalculationPrice()) {
             if ($this->hasCustomPrice()) {
-                $price = (double)$this->getCustomPrice();
+                $price = (float)$this->getCustomPrice();
                 if ($price) {
                     $rate = $this->priceCurrency->convert($price, $this->getStore()) / $price;
                     $price = $price / $rate;
@@ -489,7 +489,7 @@ abstract class AbstractItem extends \Magento\Framework\Model\AbstractExtensibleM
     {
         if (!$this->hasBaseCalculationPrice()) {
             if ($this->hasOriginalCustomPrice()) {
-                $price = (double)$this->getOriginalCustomPrice();
+                $price = (float)$this->getOriginalCustomPrice();
                 if ($price) {
                     $rate = $this->priceCurrency->convert($price, $this->getStore()) / $price;
                     $price = $price / $rate;
