+++ app/code/Magento/Sales/Model/Order/Payment/Operations/CaptureOperation.php
@@ -84,7 +84,7 @@ class CaptureOperation extends AbstractOperation
         // prepare parent transaction and its amount
         $paidWorkaround = 0;
         if (!$invoice->wasPayCalled()) {
-            $paidWorkaround = (double)$amountToCapture;
+            $paidWorkaround = (float)$amountToCapture;
         }
         if ($payment->isCaptureFinal($paidWorkaround)) {
             $payment->setShouldCloseParentTransaction(true);
